move shared components to files
