--- src/sections/Projects/Sistent/components/select/header.js
@@ -1,13 +1,0 @@
-import React from "react";
-export default function Header() {
-  return (
-    <>
-      <a id="Identity">
-        <h2>Select</h2>
-      </a>
-      <p>
-        Select component is a dropdown menu for selecting an option from a list.
-      </p>
-    </>
-  );
-}
