+++ src/sections/Projects/Sistent/components/select/index.js
@@ -2,7 +2,10 @@ import React from "react";
 import { SistentThemeProvider } from "@layer5/sistent";
 import { SistentLayout } from "../../sistent-layout";
 import { Row } from "../../../../../reusecore/Layout";
+import { useLocation } from "@reach/router";
 import { useStyledDarkMode } from "../../../../../theme/app/useStyledDarkMode";
+import { navigate } from "gatsby";
+import TabButton from "../../../../../reusecore/Button";
 import {
   FormControl,
   InputLabel,
@@ -10,8 +13,6 @@ import {
   Select,
   OutlinedInput,
 } from "@mui/material";
-import Header from "./header";
-import SectionNav from "./section-nav";
 
 const SistentSelect = () => {
   const { isDark } = useStyledDarkMode();
@@ -222,5 +223,53 @@ const SistentSelect = () => {
     </SistentLayout>
   );
 };
+
+export function SectionNav() {
+  const location = useLocation();
+  return (
+    <div className="filterBtns">
+      <TabButton
+        className={
+          location.pathname === "/projects/sistent/components/select"
+            ? "active"
+            : ""
+        }
+        onClick={() => navigate("/projects/sistent/components/select")}
+        title="Overview"
+      />
+      <TabButton
+        className={
+          location.pathname === "/projects/sistent/components/selet/guidance"
+            ? "active"
+            : ""
+        }
+        onClick={() => navigate("/projects/sistent/components/select/guidance")}
+        title="Guidance"
+      />
+      <TabButton
+        className={
+          location.pathname === "/projects/sistent/components/select/code"
+            ? "active"
+            : ""
+        }
+        onClick={() => navigate("/projects/sistent/components/select/code")}
+        title="Code"
+      />
+    </div>
+  );
+}
+
+export function Header() {
+  return (
+    <>
+      <a id="Identity">
+        <h2>Select</h2>
+      </a>
+      <p>
+        Select component is a dropdown menu for selecting an option from a list.
+      </p>
+    </>
+  );
+}
 
 export default SistentSelect;
